Revert "move around source files to compile on arduino, (not compiling yet, but still works in PIO"
This reverts commit 66b7ec96b2d00c78c572f4088980267880817d16.
