+++ src/controller.cpp
@@ -1,6 +1,5 @@
 #include "state_machine/states/InitializeState.h"
 #include "controller.h"
-#include "debugger.h"
 
 Controller::Controller() :
     Context(),
@@ -18,15 +17,15 @@ void Controller::setup_state_machine() {
 void Controller::initialize() {
   _mode_button.set_observer(this);
   _config.set_all();
-  schedule_event(Event_enum::e_controller_initialized);
+  schedue_event(Event_enum::e_controller_initialized);
 }
 
 void Controller::on_button_pressed(Button* button, uint32_t millis) {
   if(button->get_pin() == MODE_BUTTON_PIN) {
     if (millis > 4000)
-      schedule_event(Event_enum::e_button_long_pressed);
+      schedue_event(Event_enum::e_button_long_pressed);
     else
-      schedule_event(Event_enum::e_button_pressed);
+      schedue_event(Event_enum::e_button_pressed);
   }
 }
