+++ src/state_machine/abstract_state.h
