+++ src/state_machine/context.h
@@ -4,7 +4,7 @@
 #define MAX_EVENTS 10
 
 #include "abstract_state.h"
-#include "../utils/circular_buffer.h"
+#include <circular_buffer.h>
 
 
 /**
@@ -30,7 +30,7 @@ class Context {
    * Schedule a new event
    * @param event_id
    */
-  void schedule_event(Event_enum event_id);
+  void schedue_event(Event_enum event_id);
  private:
 
   AbstractState* currentState;
