+++ src/state_machine/states/state.h
@@ -1,7 +1,8 @@
 #ifndef BGEIGIE_POINTCAST_STATE_HPP
 #define BGEIGIE_POINTCAST_STATE_HPP
 
-#include "../../controller.h"
+#include <debugger.h>
+#include "controller.h"
 
 typedef enum {
   e_initializing,
